Extract validation code from node/src/lib.rs
No semantic changes. This patch is pure refactoring.

See #391
